requirements.txt
kafka_python==1.3.1
tweepy==3.5.0
kafka==1.3.1
msgpack_python==0.4.8
lz4==0.8.2
lz4tools==1.3.1.2
xxhash==0.6.1
